requirements.txt
numpy==1.24.3
opencv-python-headless==4.8.0.76
scikit-learn==1.3.0
Pillow==10.0.0 